@@ -26,58 +26,27 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Scanner;
 
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
+
 /**
  * These utilities will be used to communicate with the network.
  */
 public class NetworkUtils {
 
     final static String BASE_URL =
-            "http://api.themoviedb.org/3/movie";
+            "http://api.themoviedb.org/3/movie/";
     final public static String BASE_IMAGE_URL="http://image.tmdb.org/t/p/w780";
-    //TODO add yout api_key to test
-    final static String API_KEY="";
-    final static String API_KEY_QUERY="api_key";
-
-
-    public static URL buildUrl(String sortBy) {
-        Uri builtUri = Uri.parse(BASE_URL).buildUpon()
-                .appendPath(sortBy)
-                .appendQueryParameter(API_KEY_QUERY, API_KEY)
+    //TODO add your api_key to test
+    public static final String API_KEY="";
+    public static final  String POPULAR_END_POINT = "popular";
+    public static final String TOP_RATED_END_POINT = "top_rated";
+    public static final String FAVORITES = "favorites";
+    public static Retrofit getClient(){
+        Retrofit retrofit = new Retrofit.Builder()
+                .baseUrl(BASE_URL)
+                .addConverterFactory(GsonConverterFactory.create())
                 .build();
-
-        URL url = null;
-        try {
-            url = new URL(builtUri.toString());
-            Log.wtf("urlGenerated",url.toString());
-        } catch (MalformedURLException e) {
-            e.printStackTrace();
-        }
-        return url;
-    }
-
-    /**
-     * This method returns the entire result from the HTTP response.
-     *
-     * @param url The URL to fetch the HTTP response from.
-     * @return The contents of the HTTP response.
-     * @throws IOException Related to network and stream reading
-     */
-    public static String getResponseFromHttpUrl(URL url) throws IOException {
-        HttpURLConnection urlConnection = (HttpURLConnection) url.openConnection();
-        try {
-            InputStream in = urlConnection.getInputStream();
-
-            Scanner scanner = new Scanner(in);
-            scanner.useDelimiter("\\A");
-
-            boolean hasInput = scanner.hasNext();
-            if (hasInput) {
-                return scanner.next();
-            } else {
-                return null;
-            }
-        } finally {
-            urlConnection.disconnect();
-        }
+        return retrofit;
     }
 }
